requirements.txt
langcodes==3.3.0
matplotlib==3.7.1
openai_whisper==20230314
streamlit==1.25.0
torch==2.0.1+cu118
